# Task3/arkpz-pzpi-22-6-harmaiev-mykhailo-task3/DTO/SensorDto.cs
namespace RoadMonitoringSystem.DTO
{
    public class SensorDto
    {
        public int RoadSectionID { get; set; } 
        public string SensorType { get; set; }
        public DateTime InstallationDate { get; set; }
        public string Status { get; set; }
    }
}