# Task3/arkpz-pzpi-22-6-harmaiev-mykhailo-task3/arkpz-pzpi-22-6-harmaiev-mykhailo-task3/DTO/RoadSectionDto.cs
namespace RoadMonitoringSystem.DTO
{
    public class RoadSectionDto
    {
        public string Name { get; set; } = string.Empty;
        public string Location { get; set; } = string.Empty;
        public DateTime CreatedDate { get; set; } = DateTime.UtcNow;
    }
}