# Task5/arkpz-pzpi-22-6-harmaiev-mykhailo-task5/RoadMonitoringSystem/DTO/UserLoginDto.cs
using System.ComponentModel.DataAnnotations;

namespace RoadMonitoringSystem.DTO
{
    public class UserLoginDto
    {
        [Required]
        [MaxLength(50)]
        public string Username { get; set; }

        [Required]
        public string Password { get; set; }
    }
}